// apps/logger/types/db.d.ts
export interface definitions {
  log_items: {
    /**
     * Format: bigint
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    created_at?: string;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    updated_at?: string;
    /** Format: text */
    title: string;
    /** Format: date */
    date: string;
    /**
     * Format: bigint
     * @description Note:
     * This is a Foreign Key to `log_groups.id`.<fk table='log_groups' column='id'/>
     */
    group: number;
    /** Format: timestamp with time zone */
    ended_at?: string;
  };
  log_groups: {
    /**
     * Format: bigint
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    created_at?: string;
    /**
     * Format: timestamp with time zone
     * @default now()
     */
    updated_at?: string;
    /** Format: text */
    title: string;
    /** Format: smallint */
    rank: number;
  };
}